+++ deploy/Strategy.js
@@ -7,16 +7,6 @@ module.exports = async ({getNamedAccounts, deployments}) => {
         log: true,
     });
 
-    await deploy('StrategyKeeperDao', {
-        from: deployer,
-        log: true,
-    });
-
-    await deploy('StrategyMStable', {
-        from: deployer,
-        log: true,
-    });
-
     await deploy('StrategyYearnAlusd', {
         from: deployer,
         log: true,